refactor: PRレビュー指摘の対応
- MaskingImageItem と createDownloadFileName を types.ts に集約し、MaskingGallery / GalleryItem の重複定義を削除
- 画像の arrayBuffer 変換を Promise.allSettled に変更し、失敗時もエラーハンドリングを実施
- 変換完了した画像を 1 枚ずつ即時 setImages して早期表示
- isProcessing 中も FaceDetectionCanvas を表示しオーバーレイで処理中インジケータを表示
- next.config.ts の条件分岐スプレッドを ternary 形式に修正して型エラーを解消
